2nd_homework/requirements.txt
Package      Version
------------ -------
pip          21.3.1
pycryptodome 3.19.0
setuptools   58.5.3
wheel        0.37.0
